requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
motor
httpx
pydantic
python-dotenv
django>=5.0.0
passlib[bcrypt]
pyjwt
email-validator
python-multipart
PyPDF2
python-docx
openpyxl
huggingface-hub
transformers
torch
sentencepiece
